[STRATO-1766] Minor enhancement
